+++ source/style/twitterFirstMinute.js
@@ -1,11 +1,12 @@
 import { style } from './twitter'
 
-// Skip "seconds", "now", "seconds" (from `0.5`) steps.
-const steps = style.steps.slice(3)
+// Skip "seconds".
+const steps = style.steps.slice(1)
+
+// Starts showing `1m` after 40 seconds.
 steps[0] = {
 	...steps[0],
-	// Starts showing `1m` after 45.5 seconds.
-	minTime: 45.5
+	minTime: 40
 }
 
 export default {